requirements.txt
flask>=2.3.0
flask-cors>=4.0.0
gunicorn>=21.0.0
cryptography>=41.0.0
requests>=2.31.0
psycopg2-binary>=2.9.0
